Podcast Hero Banner
Apologies. I ended up ruining the PR #105 with the wrong rebase, I am in the process of setting up a new desktop and got things mucked up.

I have now done a new PR with the changes and feedback from the first PR, apologies for this miss... error on my part