Adds templates for Complex blog (News)
Adds templates for Complex blog (News) listed in #42 named as following:

- Blog home with various grids
- Post with sidebar
- News blog archive
- News blog search results

I could not make the home layout work in all screen sizes with the grid block, I used good old columns block. If you find a way to make it work and be responsive, I'd love to learn how.