Remove the priority on the init hook and fix the typo in the translation function
It is not documented why the priority on the init hook is needed, and removing it does not break anything, so, it should be removed.

The translation function for the copyright text was accidently added as only an escaping function. By changing esc_html() to esc_html()__ the text is both translated and escaped, which is the desired outcome.