News blog with sidebar: Add main element
This commit wraps the top section and the bottom section in one group instead of using two gorups.
The tag on the group is changed from div to main.
A spacer block with spacing preset 50 is placed between the two sections, to replace the padding that was on the group that was removed.